+++ app/controllers/bookingsync_portal/admin/rentals_controller.rb
@@ -16,10 +16,22 @@ module BookingsyncPortal
         rental
       end
 
+      def connect_to_new
+        unless BookingsyncPortal.create_remote_rental_from_app
+          raise ActionController::RoutingError.new('Not Found')
+        end
+
+        remote_account = current_account.remote_accounts.find(params[:remote_account_id])
+        remote_rental = ::RemoteRental.create!(remote_account_id: remote_account.id)
+        connection = rental.create_connection(remote_rental: remote_rental)
+
+        BookingsyncPortal.remote_rental_created(remote_rental)
+        redirect_or_js_response
+      end
+
       def connect
         remote_rental = current_account.remote_rentals.find(params[:remote_rental_id])
-        connection = rental.build_connection(remote_rental: remote_rental)
-        connection.save
+        connection = rental.create_connection(remote_rental: remote_rental)
 
         BookingsyncPortal.connection_created(connection)
         redirect_or_js_response