+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Remove CircleCI setup
 * Cleanup bookingsync_portal config generator and dummy app one
 * Add missing dummy simple_form initializer
+* Add `BookingsyncPortal.remote_rental_created` class method.
+* Add `BookingsyncPortal.create_remote_rental_from_app` option to allow to create remote rentals directly from the app. Default to `false`.
 
 # 0.6.0